AssetManager::reload returns size_t

+++ include/GameLibrary/Application/AssetManager.hpp
@@ -101,7 +101,7 @@ namespace fgl
 		
 		/*! Reloads and re-stores all stored assets from their paths.
 			\returns the total amount of successfully reloaded assets*/
-		unsigned int reload();
+		size_t reload();
 		/*! Unloads and deallocates all stored assets. This does not include assets stored in dependent asset managers */
 		void unload();
